Fix Timestamp Formatting
Timestamps were previously exported like so:

    20200902T02031Z

Note however that the time component is expressed as '02031Z', which
oddly includes the tens digits of the seconds but not the ones digit,
making this timestamp unparseable automatically by the likes of Ruby's
`Time.parse`.

This edit modifies the regex to not capture the seconds ones digit,
causing the following to be generated as the timezone.

    20200902T020319Z